Auth: add server-side redirect, simplify DynamicImage, consolidate GoogleAuthButton
- layout.tsx: server component now calls getServerSession() and redirects
  to /document if already authenticated, preventing flash on auth pages
- DynamicImage.tsx: replace useCallback/useEffect/useState with a single
  static lookup object; derive image directly from pathname
- GoogleAuthButton: move duplicate signin/signup copies to shared
  app/(auth)/components/ and update both page imports
